menu is in a separate cpp file now and matrix.h functions are in alphabetical order.
