+++ Matrix/makefile
@@ -1,4 +1,4 @@
 Matrix: atest.cpp matrix.cpp functionality.cpp matrix.h
-	g++ -std=c++11 -g -o Matrix atest.cpp matrix.cpp functionality.cpp
+	g++ -std=c++11 -g -o Matrix atest.cpp matrix.cpp functionality.cpp menu.cpp
 clean:
 	rm -f Matrix
