+++ Matrix/matrix.h
@@ -17,58 +17,59 @@ public:
     ~Matrix();
     // These functions are all useful matrix operations.
     Matrix add(Matrix Other);
+    Matrix cross(Matrix Other, bool keep);
     double characteristic();
-    double determinant(vector<vector<double> > toDetermine, int column, int row);
+    double decEntry();
+    double determinant(vector<int> rem);
     void eigenValues();
     void eigenVecs();
     void eigen1x1();
     void eigen2x2();
     Matrix invert();
     bool invertible();
-    Matrix cross(Matrix Other, bool keep);
     double l2norm();
-    void QR(int n, Matrix Ea);
     Matrix scale(double constant);
     void solve(Matrix solution);
     Matrix subtract(Matrix Other);
     double trace();
     Matrix transpose();
-    // These are more of overhead functions for the program.
+    // The matrix menu is in its own file so it can be easily accessible.
     void menu();
-    double determinant();
-    void publishFile();
-    vector<double> polynomial(int row);
-    bool checkCompatibility(Matrix Other);
-    int vectorPosition(vector<Matrix> matrixMenu, vector<int> numMenu);
-    void checkValidity(ifstream & readMatrix);
-    bool ynResponse();
-    int getWidth();
-    int getHeight();
-    void switchRows(vector<vector<double> > toSwap, int row1, int row2);
-    vector<vector<double> > getMatrix();
-    void printMatrix();
-    void publishNew(int rows, int columns, vector<vector<double> > newMatrix);
-    void printEigen();
-    string generateName();
-    string findName();
-    bool square();
-    vector<vector<double> > rref();
-    string getName();
+    // These are more of overhead functions for the program.
     void addToList(Matrix A);
-    Matrix getMatrix(string number);
-    string longName(string shortName);
-    Matrix householder();
-    void setMatrix(vector<vector<double> > toSet);
-    vector<vector<double> > wilkinson(double a, double b, double c);
-    vector<vector<double> > identity(double factor);
+    bool checkCompatibility(Matrix Other);
+    void checkValidity(ifstream & readMatrix);
     bool compareQR();
-    void printEigenVec();
-    int strChoice();
+    double determinant();
+    string findName();
+    string generateName();
+    int getHeight();
+    vector<vector<double> > getMatrix();
+    string getName();
+    int getWidth();
+    Matrix householder();
+    vector<vector<double> > identity(double factor);
     int intEntry();
-    double decEntry();
-    vector<vector<double> > returnEigen();
-    vector<vector<double> > normalizeBottom(vector<double> real, vector<double> imag);
+    string longName(string shortName);
     vector<vector<double> > multiplicity(vector<vector<double> > BaaB, int n);
+    vector<vector<double> > normalizeBottom(vector<double> real, vector<double> imag);
+    void printEigen();
+    void printEigenVec();
+    void printMatrix();
+    void publishFile();
+    void publishNew(int rows, int columns, vector<vector<double> > newMatrix);
+    vector<double> polynomial(int row);
+    void QR(int n, Matrix Ea);
+    vector<vector<double> > returnEigen();
+    Matrix retreiveMatrix(string number);
+    vector<vector<double> > rref();
+    void setMatrix(vector<vector<double> > toSet);
+    bool square();
+    int strChoice();
+    void switchRows(vector<vector<double> > toSwap, int row1, int row2);
+    int vectorPosition(vector<Matrix> matrixMenu, vector<int> numMenu);
+    vector<vector<double> > wilkinson(double a, double b, double c);
+    bool ynResponse();
 private:
     vector<vector<double> > matrix;
     string name;
